Wait for server to process upload
Now using normal async socket, in order to avoid manual buffer copy.
Also it is mandatory to wait for the response from the server that it
received the file. Previously this was marked as optional and was not
actually waiting for some response bytes. Therefore it was then possible
to cut the connection before the server processed all the data from the
data socket.